Make unused untrack_message do what is actually expected

+++ bot/submission/submit.py
@@ -28,7 +28,7 @@ from database import DatabaseManager
 from database.enums import Status, Category
 from bot._types import GuildMessageable
 from bot.utils import RunningMessage, is_owner_server, check_is_staff, check_is_trusted_or_staff
-from database.message import get_build_id_by_message
+from database.message import get_build_id_by_message, untrack_message
 from database.schema import TypeRecord
 from database.server_settings import get_server_setting
 from database.utils import get_version_string, upload_to_catbox
@@ -598,14 +598,8 @@ class BuildCog(Cog, name="Build"):
     @edit_group.command(name="door")
     async def edit_door(self, ctx: Context, *, flags: EditDoorFlags):
         """Edits a door record in the database directly."""
-        interaction = ctx.interaction
-        assert interaction is not None
-
-        response: InteractionResponse = interaction.response  # type: ignore
-        await response.defer()
-
-        followup: discord.Webhook = interaction.followup  # type: ignore
-        async with RunningMessage(followup) as sent_message:
+        await ctx.defer()
+        async with RunningMessage(ctx) as sent_message:
             build = await flags.to_build()
             if build is None:
                 error_embed = utils.error_embed("Error", "No build with that ID.")
@@ -615,20 +609,31 @@ class BuildCog(Cog, name="Build"):
 
             # Show a preview of the changes and ask for confirmation
             await sent_message.edit(embed=utils.info_embed("Waiting", "User confirming changes..."))
-            view = ConfirmationView()
-            preview = await followup.send(embed=preview_embed, view=view, ephemeral=True, wait=True)
-            await view.wait()
-
-            await preview.delete()
-            if view.value is None:
-                await sent_message.edit(embed=utils.info_embed("Timed out", "Build edit canceled due to inactivity."))
-            elif view.value:
+            if ctx.interaction:
+                view = ConfirmationView()
+                preview = await ctx.interaction.followup.send(
+                    "Here is a preview of the changes. Use the buttons to confirm or cancel.",
+                    embed=preview_embed,
+                    view=view,
+                    ephemeral=True,
+                    wait=True
+                )
+                await view.wait()
+                await preview.delete()
+                if view.value is None:
+                    await sent_message.edit(embed=utils.info_embed("Timed out", "Build edit canceled due to inactivity."))
+                elif view.value:
+                    await sent_message.edit(embed=utils.info_embed("Editing", "Editing build..."))
+                    await build.save()
+                    await self.update_build_messages(build)
+                    await sent_message.edit(embed=utils.info_embed("Success", "Build edited successfully"))
+                else:
+                    await sent_message.edit(embed=utils.info_embed("Cancelled", "Build edit canceled by user"))
+            else:  # Not an interaction, so we can't use buttons for confirmation
                 await sent_message.edit(embed=utils.info_embed("Editing", "Editing build..."))
                 await build.save()
                 await self.update_build_messages(build)
                 await sent_message.edit(embed=utils.info_embed("Success", "Build edited successfully"))
-            else:
-                await sent_message.edit(embed=utils.info_embed("Cancelled", "Build edit canceled by user"))
 
     async def update_build_message(self, build: Build, channel_id: int, message_id: int) -> None:
         """Updates a post according to the information given by the build."""
@@ -650,14 +655,20 @@ class BuildCog(Cog, name="Build"):
             raise ValueError("Build id is None.")
 
         # Get all messages for a build
-        messages = await msg.get_build_messages(build.id)
+        message_records = await msg.get_build_messages(build.id)
         em = build.generate_embed()
 
-        for message in messages:
-            channel = self.bot.get_channel(message["channel_id"])
+        for record in message_records:
+            channel = self.bot.get_channel(record["channel_id"])
             if not isinstance(channel, GuildMessageable):
                 raise ValueError(f"Invalid channel type for a post channel: {type(channel)}")
-            message = await channel.fetch_message(message["message_id"])
+            try:
+                message = await channel.fetch_message(record["message_id"])
+            except discord.NotFound:
+                await untrack_message(record["message_id"])
+                continue
+            except discord.Forbidden:
+                continue
             await message.edit(embed=em)
             await msg.update_message_edited_time(message.id)
